backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=12.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.26.0
ultralytics==8.2.0
pytesseract==0.3.10
easyocr==1.7.0
scikit-image>=0.21.0
scipy>=1.11.0
requests==2.31.0
PyMySQL==1.1.0
psycopg2-binary==2.9.10
argon2-cffi>=21.3.0
